MigrationTestCase base class and change logs wording
