+++ kolibri/core/auth/test/test_data_migrations.py
@@ -1,50 +1,15 @@
-from django.db import connection
-from django.db.migrations.executor import MigrationExecutor
-from django.test import TestCase
-
+from .migrationtestcase import TestMigrations
 from kolibri.core.auth.constants.role_kinds import ADMIN
 from kolibri.core.auth.models import Classroom
 from kolibri.core.auth.models import Facility
 from kolibri.core.auth.models import FacilityUser
-
-# Modified from https://www.caktusgroup.com/blog/2016/02/02/writing-unit-tests-django-migrations/
-
-
-class TestMigrations(TestCase):
-
-    migrate_from = None
-    migrate_to = None
-    app = 'kolibriauth'
-
-    def setUp(self):
-        assert self.migrate_from and self.migrate_to, \
-            "TestCase '{}' must define migrate_from and migrate_to properties".format(type(self).__name__)
-
-        migrate_from = [(self.app, self.migrate_from)]
-        migrate_to = [(self.app, self.migrate_to)]
-        executor = MigrationExecutor(connection)
-        old_apps = executor.loader.project_state(migrate_from).apps
-
-        # Reverse to the original migration
-        executor.migrate(migrate_from)
-
-        self.setUpBeforeMigration(old_apps)
-
-        # Run the migration to test
-        executor = MigrationExecutor(connection)
-        executor.loader.build_graph()  # reload.
-        executor.migrate(migrate_to)
-
-        self.apps = executor.loader.project_state(migrate_to).apps
-
-    def setUpBeforeMigration(self, apps):
-        pass
 
 
 class MultipleCollectionTestCase(TestMigrations):
 
     migrate_from = '0003_auto_20170621_0958'
     migrate_to = '0004_auto_20170816_1607'
+    app = 'kolibriauth'
 
     def setUp(self):
         self.facility = Facility.objects.create(name='Test')
