+++ kolibri/core/content/test/test_data_migrations.py
@@ -1,51 +1,15 @@
 import uuid
 
-from django.db import connection
-from django.db.migrations.executor import MigrationExecutor
-from django.test import TestCase
-
+from kolibri.core.auth.test.migrationtestcase import TestMigrations
 from kolibri.core.content.models import ChannelMetadata as RealChannelMetadata
 from kolibri.core.content.models import ContentNode as RealContentNode
-
-
-# Modified from https://www.caktusgroup.com/blog/2016/02/02/writing-unit-tests-django-migrations/
-
-
-class TestMigrations(TestCase):
-
-    migrate_from = None
-    migrate_to = None
-    app = 'content'
-
-    def setUp(self):
-        assert self.migrate_from and self.migrate_to, \
-            "TestCase '{}' must define migrate_from and migrate_to properties".format(type(self).__name__)
-
-        migrate_from = [(self.app, self.migrate_from)]
-        migrate_to = [(self.app, self.migrate_to)]
-        executor = MigrationExecutor(connection)
-        old_apps = executor.loader.project_state(migrate_from).apps
-
-        # Reverse to the original migration
-        executor.migrate(migrate_from)
-
-        self.setUpBeforeMigration(old_apps)
-
-        # Run the migration to test
-        executor = MigrationExecutor(connection)
-        executor.loader.build_graph()  # reload.
-        executor.migrate(migrate_to)
-
-        self.apps = executor.loader.project_state(migrate_to).apps
-
-    def setUpBeforeMigration(self, apps):
-        pass
 
 
 class ChannelFieldsTestCase(TestMigrations):
 
     migrate_from = '0011_auto_20180907_1017'
     migrate_to = '0012_auto_20180910_1702'
+    app = 'content'
 
     def setUp(self):
         self.file_size = 10
